+++ designsafe/apps/search/static/designsafe/apps/search/scripts/app.js
@@ -1,8 +1,10 @@
 (function(window, angular, $, _) {
   "use strict";
-
-  angular.module('designsafe')
-  .config(["$httpProvider", "$locationProvider", function ($httpProvider, $locationProvider) {
+  var module = angular.module('designsafe');
+  module.requires.push(
+    'ngSanitize'
+  );
+  module.config(["$httpProvider", "$locationProvider", function ($httpProvider, $locationProvider) {
      $httpProvider.defaults.transformResponse.push(function(responseData){
         convertDateStringsToDates(responseData);
         return responseData;
@@ -13,9 +15,10 @@
        });
   }]);
 
-  angular.module('designsafe').controller('SearchCtrl',
-    ['$scope','$rootScope', '$location', '$window', 'searchService', 'Logging', 'djangoUrl',
-    function($scope, $rootScope, $location, $window, searchService, Logging, djangoUrl) {
+  module.controller('SearchCtrl',
+    ['$scope','$rootScope', '$location', '$window', 'searchService', 'Logging', 'djangoUrl', 'Django',
+    function($scope, $rootScope, $location, $window, searchService, Logging, djangoUrl, Django) {
+      $scope.user = Django.user;
       $scope.data = {};
       $scope.Math = window.Math;
       $scope.counter = Array;
@@ -23,7 +26,7 @@
       $scope.results_per_page = 10;
       $scope.offset = 0;
       $scope.data.search_text = null;
-      $scope.data.type_filter = null;
+      $scope.data.type_filter = 'cms';
       $scope.filetype_filter = 'all';
       $scope.searching = false;
       $scope.inital_q = $location.search().q;
@@ -50,12 +53,7 @@
       };
 
       $scope.filter = function (ftype) {
-        // toggle the
-        if ($scope.data.type_filter === ftype) {
-          $scope.data.type_filter = null;
-        } else {
-          $scope.data.type_filter = ftype;
-        }
+        $scope.data.type_filter = ftype;
         $scope.page_num = 0;
         $scope.search();
       };
@@ -97,7 +95,7 @@
         var match;
         // Check for string properties which look like dates.
         if (typeof value === "string" && (match = value.match(regexIso8601))) {
-            var milliseconds = Date.parse(match[0])
+            var milliseconds = Date.parse(match[0]);
             if (!isNaN(milliseconds)) {
                 input[key] = new Date(milliseconds);
             }